+++ src/openforms/submissions/form_logic.py
@@ -1,11 +1,8 @@
 from typing import TYPE_CHECKING, Any, Dict, List, Optional
-from urllib.parse import quote
 
 from django.utils.functional import empty
 
 import elasticapm
-import jq
-from json_logic import jsonLogic
 
 from openforms.formio.service import (
     get_dynamic_configuration,
@@ -13,11 +10,7 @@ from openforms.formio.service import (
     translate_function,
 )
 from openforms.formio.utils import get_component_empty_value, is_visible_in_frontend
-from openforms.forms.models import FormVariable
 from openforms.logging import logevent
-from openforms.typing import DataMapping, JSONValue
-from openforms.variables.models import DataMappingTypes, ServiceFetchConfiguration
-from openforms.variables.validators import HeaderValidator
 
 from .logic.actions import PropertyAction
 from .logic.datastructures import DataContainer
@@ -255,61 +248,3 @@ def check_submission_logic(
         mutation.apply(step, {})
 
     submission._form_logic_evaluated = True
-
-
-def bind(var: FormVariable, context: DataMapping) -> JSONValue:
-    """Bind a value to a variable `var`
-
-    :raises: :class:`requests.HTTPException` for internal server errors
-    :raises: :class:`zds_client.client.ClientError` for HTTP 4xx status codes
-    :raises: :class:`NotImplementedError` for FormVariable
-    """
-    fetch_config: ServiceFetchConfiguration | None
-
-    match var:
-        case FormVariable(service_fetch_configuration=fetch_config) if fetch_config:
-            client = fetch_config.service.build_client()
-            request_args = _request_arguments(fetch_config, context)
-            raw_value = client.request(**request_args)
-
-            match fetch_config.data_mapping_type, fetch_config.mapping_expression:
-                case DataMappingTypes.jq, expression:
-                    # XXX raise warning if len(result) > 1 ?
-                    value = jq.compile(expression).input(raw_value).first()
-                case DataMappingTypes.json_logic, expression:
-                    value = jsonLogic(expression, raw_value)
-                case _:
-                    value = raw_value
-        case _:
-            raise NotImplementedError()
-
-    return value
-
-
-def _request_arguments(fetch_config: ServiceFetchConfiguration, context: DataMapping):
-    headers = {
-        # no leading spaces and the value into the legal field-value codespace
-        header: value.format(**context).strip().encode("utf-8").decode("latin1")
-        for header, value in (fetch_config.headers or {}).items()
-    }
-    # before we go further
-    # assert headers are still valid after we put submmitter data in there
-    HeaderValidator()(headers)
-
-    escaped_values = {k: quote(str(v), safe="") for k, v in context.items()}
-
-    query_params = fetch_config.query_params.format(**escaped_values)
-    query_params = query_params[1:] if query_params.startswith("?") else query_params
-
-    request_args = dict(
-        path=fetch_config.path.format(**escaped_values),
-        params=query_params,
-        method=fetch_config.method,
-        headers=headers,
-        # XXX operationId is a required parameter to zds-client...
-        # Maybe offer as alternative to path and method?
-        operation="",
-    )
-    if fetch_config.body is not None:
-        request_args["json"] = fetch_config.body
-    return request_args
